Consolidate all types inside the module
