+++ src/test/api.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import { expect } from 'chai';
-import { PythonExtension } from '../client/api/apiTypes';
+import { PythonExtension } from '../client/api/main';
 import { ProposedExtensionAPI } from '../client/proposedApiTypes';
 import { initialize } from './initialize';
